+++ docs/Changelog.md
@@ -8,6 +8,7 @@
 - Color filters in map rendering [#778](https://github.com/mapsforge/mapsforge/issues/778)
 - Night mode example [#777](https://github.com/mapsforge/mapsforge/issues/777)
 - Zoom level limits central handling [#646](https://github.com/mapsforge/mapsforge/issues/646)
+- Android gestures options [#705](https://github.com/mapsforge/mapsforge/issues/705)
 - Java onTap layer listener [#774](https://github.com/mapsforge/mapsforge/issues/774)
 - Library integration [documentation](Integration.md) [#781](https://github.com/mapsforge/mapsforge/issues/781)
 - Jar with dependencies building [#767](https://github.com/mapsforge/mapsforge/issues/767)